Bug fix, remove old comments, and refactor functions
This is the second half of the refactoring process started in the previous commit.
It is now possible for ingredients (not just coffees) to use the buildDescriptionFor function.
However, this required a few related functions to be changed.

As for the bug, the previous commit accidentally caused an ingredient's amount to be displayed twice.
This has been fixed.