Add heading and form to index.html, as a default
My plan is to have the page load the "New Concoction" form and its heading by default.
In order to enforce Separation of Concerns, I am moving this behavior from the index.js file to the index.html file.
The next commit should remove the JS for creating a "New Concoction" form.