Combine two Ingredient methods into one
I don't know if this is the best way to do it, but I'm trying to obey Separation of Concerns here.
Even though I'd be able to use an instance method and the 'this' keyword,
it doesn't make sense to me to separate 'adding an ingredient to a list' into another method.