Adjust padding, margin, height, and flex/grid styles
This improves the overall look of the app. The grid gaps are constant,
and the #wrapper grid finally fits its contents instead of leaving large gaps of white and brown space.
The top and bottom padding are now the same, and there are minor height and spacing improvements to the
"New Concoction" button and "Saved Concoctions" list.